+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [0.3.3]
 - Remove sleep from updater (request only)
+- Raise errors on auth and update failure
 
 ## [0.3.2]
 - Add water heater set temperature service